Add foss build flavor
