+++ mobile/src/foss/java/org/openhab/habdroid/core/GcmIntentService.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2010-2016, openHAB.org and others.
+ *
+ *   All rights reserved. This program and the accompanying materials
+ *   are made available under the terms of the Eclipse Public License v1.0
+ *   which accompanies this distribution, and is available at
+ *   http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.openhab.habdroid.core;
+
+import android.app.IntentService;
+import android.content.Intent;
+
+public class GcmIntentService extends IntentService {
+    public GcmIntentService() {
+        super("GcmIntentService");
+    }
+
+    @Override
+    protected void onHandleIntent(Intent intent) {
+    }
+}
