+++ mobile/src/foss/java/org/openhab/habdroid/core/notifications/GoogleCloudMessageConnector.java
@@ -1,0 +1,9 @@
+package org.openhab.habdroid.core.notifications;
+
+public class GoogleCloudMessageConnector {
+    private static final String TAG = GoogleCloudMessageConnector.class.getSimpleName();
+
+    public boolean register() {
+        return false;
+    }
+}
