+++ mobile/src/foss/java/org/openhab/habdroid/nonfree/NonfreeMainActivity.java
@@ -1,0 +1,14 @@
+package org.openhab.habdroid.nonfree;
+
+import android.content.Intent;
+import android.os.AsyncTask;
+import android.provider.Settings;
+
+import org.openhab.habdroid.ui.OpenHABMainActivity;
+
+public class NonfreeMainActivity extends OpenHABMainActivity {
+    public void gcmRegisterBackground() {}
+
+    public void processIntent(Intent intent) {
+    }
+}
