+++ mobile/src/foss/java/org/openhab/habdroid/nonfree/myGoogleCloudMessaging.java
@@ -1,0 +1,9 @@
+package org.openhab.habdroid.nonfree;
+
+import android.content.Context;
+
+public class myGoogleCloudMessaging {
+    public static myGoogleCloudMessaging getInstance(Context context) {
+        return null;
+    }
+}
