+++ mobile/src/full/java/org/openhab/habdroid/core/NotificationDeletedBroadcastReceiver.java
