+++ mobile/src/full/java/org/openhab/habdroid/core/notifications/GoogleCloudMessageConnector.java
