+++ mobile/src/full/java/org/openhab/habdroid/nonfree/myGoogleCloudMessaging.java
@@ -1,0 +1,6 @@
+package org.openhab.habdroid.nonfree;
+
+import com.google.android.gms.gcm.GoogleCloudMessaging;
+
+public class myGoogleCloudMessaging extends GoogleCloudMessaging {
+}
