+++ mobile/src/main/AndroidManifest.xml
@@ -75,6 +75,7 @@
         <activity android:name="org.openhab.habdroid.ui.IntroActivity"
             android:label="@string/app_intro"/>
         <activity android:name="de.duenndns.ssl.MemorizingActivity" />
+        <activity android:name=".nonfree.NonfreeMainActivity" />
 
         <service
             android:name=".core.OpenHABVoiceService"
